% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACO.R
\docType{class}
\name{ACO-class}
\alias{ACO-class}
\title{An S4 class for the Ant Colony Optimization Algorithm}
\value{
An S4 object of class `ACO`.
}
\description{
An S4 class for the Ant Colony Optimization Algorithm
}
\section{Slots}{

\describe{
\item{\code{function_call}}{The original function call.}

\item{\code{summary}}{A summary `data.frame` indicating the algorithm results for
each iteration.}

\item{\code{final_solution}}{A `matrix` with the final solution information, including
fit indices, selected items, and pheromone level.}

\item{\code{best_model}}{A `lavaan` object of the final solution.}

\item{\code{best_syntax}}{A `character` vector of the final solution model syntax.}

\item{\code{runtime}}{A `difftime` object of the total run time of the function.}
}}

