% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antcolony_plot.R
\name{antcolony_plot}
\alias{antcolony_plot}
\title{ACO Plot Function}
\usage{
antcolony_plot(results, ...)
}
\arguments{
\item{results}{The output from `antcolony.lavaan`.}

\item{...}{Other arguments to be passed to the `ggplot2` functions. Currently
unimplemented.}
}
\value{
A list of three plots: 
\describe{
\item{1}{A pheromone plot, showing how the pheromone levels for each item change 
for each run of the algorithm;}
\item{2}{A mean gamma plot, showing a loess smoothing plot of changes in the mean
gamma (regression coefficients) of the best-fit model for each run of the 
algorithm;}
\item{3}{A mean variance explained plot, showing a loess smoothing plot of 
changes in the mean variance explained of the best-fit model for each run of the 
algorithm}
}
}
\description{
ACO Plot Function
}
