% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleAntModel}
\alias{exampleAntModel}
\title{Model syntax for the example in the \code{\link{antColony}} function.}
\format{
A character vector.
}
\usage{
data(exampleAntModel)
}
\description{
A character vector containing the model syntax used for the one factor, 56-item example in the \code{\link{antColony}}.
}
\keyword{datasets}
