% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelCheck.R
\docType{class}
\name{modelCheck-class}
\alias{modelCheck-class}
\title{An S4 class for the modelCheck object}
\value{
An S4 object of class `ACO`.
}
\description{
An S4 class for the modelCheck object
}
\section{Slots}{

\describe{
\item{\code{model.output}}{A `lavaan` object.}

\item{\code{warnings}}{A `character` vector of any warnings.}

\item{\code{errors}}{A `character` vector of any errors.}

\item{\code{model.syntax}}{A `character` vector of the modelCheck model syntax.}
}}

