% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabu_internal.R
\name{par.matches}
\alias{par.matches}
\title{A series of internal functions used by the Tabu search.
Directly accessing these functions is not advised unless you know what you are doing.}
\usage{
par.matches(ptab, label, block = NULL)
}
\description{
A series of internal functions used by the Tabu search.
Directly accessing these functions is not advised unless you know what you are doing.
}
\keyword{internal}
