% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACO.R
\name{plot,ACO,ANY-method}
\alias{plot,ACO,ANY-method}
\title{Plot method for class `ACO`}
\usage{
\S4method{plot}{ACO,ANY}(x, y, type = "all", ...)
}
\arguments{
\item{x, y}{An S4 object of class `ACO`}

\item{type}{A `character` value specifying the plot type. One of `'all'` 
(for all plots), `'pheromone'`, `'gamma'`, `'beta'`, or `'variance'`.}

\item{...}{Not used.}
}
\description{
Plot method for class `ACO`
}
