% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{plot,TS,ANY-method}
\alias{plot,TS,ANY-method}
\title{Plot method for class `TS`}
\usage{
\S4method{plot}{TS,ANY}(x, y, ...)
}
\arguments{
\item{x, y}{An S4 object of class `TS`. `y` is included for method
compatibility and is not used.}

\item{...}{Not used.}
}
\description{
Plots the objective function's value across iterations.
}
