% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabu_refit.R
\name{refit.model}
\alias{refit.model}
\title{Given a fitted lavaan model and a search table, refits the model using the search
table as specifying what changes should be done (parmeters fixed/freed).}
\usage{
refit.model(fitted.model, ptab)
}
\arguments{
\item{fitted.model}{fitted model of class lavaan}

\item{ptab}{search table}
}
\value{
An object of class \code{lavaan} if the new model fits, or an object of class \code{try-error} if the model update fails.
}
\description{
This is not meant to be called explicity as \link{tabu.sem} uses this
internally for model refitting.
}
\references{
\url{https://doi.org/10.1080/10705511.2017.1409074}
}
\seealso{
Other Tabu Search: \code{\link{add.param}},
  \code{\link{search.prep}}
}
\author{
Carl F. Falk
}
\concept{Tabu Search}
