% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\docType{class}
\name{SA-class}
\alias{SA-class}
\title{An S4 class for the Simulated Annealing Algorithm}
\value{
An S4 object of class `SA`.
}
\description{
An S4 class for the Simulated Annealing Algorithm
}
\section{Slots}{

\describe{
\item{\code{function_call}}{The original function call.}

\item{\code{chains}}{The number of chains used.}

\item{\code{chain_results}}{A `matrix` (for multiple chains) or a `list` (for a single chain) of the chain results.}

\item{\code{all_fit}}{A summary `vector` indicating the model fit results for
each iteration.}

\item{\code{best_fit}}{The best model fit result using the selected `fitStatistic`.}

\item{\code{best_model}}{A `lavaan` object of the final solution.}

\item{\code{best_syntax}}{A `character` vector of the final solution model syntax.}

\item{\code{runtime}}{A `difftime` object of the total run time of the function.}
}}

