% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortExampleAntModel}
\alias{shortExampleAntModel}
\title{Model syntax for the short example in the \code{\link{antColony}} function.}
\format{
A character vector.
}
\usage{
data(shortExampleAntModel)
}
\description{
A character vector containing the model syntax used for the one factor, 15-item, example in the \code{\link{antColony}}.
}
\keyword{datasets}
