% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortForm.R
\name{ShortFormStartup}
\alias{ShortFormStartup}
\title{Create Package Startup Message}
\usage{
ShortFormStartup()
}
\description{
Makes package startup message.
}
\details{
Idea taken from https://github.com/ntguardian/MCHT/blob/master/R/StartupMessage.R
}
\examples{
ShortForm:::ShortFormStartup()
}
