% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\name{show,SA-method}
\alias{show,SA-method}
\title{Print method for class `SA`}
\usage{
\S4method{show}{SA}(object)
}
\arguments{
\item{object}{An S4 object of class `SA`.}
}
\description{
Print method for class `SA`
}
