% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\name{show,TS-method}
\alias{show,TS-method}
\title{Print method for class `TS`}
\usage{
\S4method{show}{TS}(object)
}
\arguments{
\item{object}{An S4 object of class `TS`.}
}
\description{
Print method for class `TS`
}
