% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_test_data}
\alias{simulated_test_data}
\title{A simulated data set based on a standardized test.}
\format{
An object of class \code{data.frame} with 1000 rows and 58 columns.
}
\usage{
simulated_test_data
}
\description{
Simulated response patterns, abilities, and outcomes based on a unidimensional state-issued standardized test.
}
\details{
@format A data frame of 1000 rows (observations) and 58 columns (variables):
 \describe{
 \item{Outcome}{a binary external criterion variable correlated with TrueAbility}
 \item{TrueAbility}{the simulated true ability parameter used to generate response patterns}
 \item{Item1-Item56}{binary reponses to items generated using the TrueAbility parameters and simulated 3PL item parameters generated from the distribution of parameters estimated from a state-issued standardized test}
 }
}
\keyword{datasets}
