% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACO.R
\docType{methods}
\name{summary,ACO-method}
\alias{summary,ACO-method}
\title{Summary method for class `ACO`}
\usage{
\S4method{summary}{ACO}(object)
}
\arguments{
\item{object}{An S4 object of class `ACO`}
}
\description{
Summary method for class `ACO`
}
