% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\docType{methods}
\name{summary,SA-method}
\alias{summary,SA-method}
\title{Summary method for class `SA`}
\usage{
\S4method{summary}{SA}(object)
}
\arguments{
\item{object}{An S4 object of class `SA`.}
}
\description{
Summary method for class `SA`
}
