% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\docType{methods}
\name{summary,TS-method}
\alias{summary,TS-method}
\title{Summary method for class `TS`}
\usage{
\S4method{summary}{TS}(object)
}
\arguments{
\item{object}{An S4 object of class `TS`.}
}
\description{
Summary method for class `TS`
}
