% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA.R
\docType{methods}
\name{plot,SA,ANY-method}
\alias{plot,SA,ANY-method}
\title{Plot method for class `SA`}
\usage{
\S4method{plot}{SA,ANY}(x, y, ...)
}
\arguments{
\item{x, y}{An S4 object of class `SA`.}

\item{...}{Not used.}
}
\description{
Plot method for class `SA`
}
