% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.R
\docType{methods}
\name{plot,TS,ANY-method}
\alias{plot,TS,ANY-method}
\title{Plot method for class `TS`}
\usage{
\S4method{plot}{TS,ANY}(x, y, ...)
}
\arguments{
\item{x, y}{An S4 object of class `TS`.}

\item{...}{Not used.}
}
\description{
Plot method for class `TS`
}
