% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShortForm.R
\name{ShortForm}
\alias{ShortForm}
\alias{ShortForm-package}
\title{\code{ShortForm} package}
\description{
Automated Item Selection Algorithms for Short Forms
}
\details{
See the README on \href{https://github.com/AnthonyRaborn/ShortForm#readme}{GitHub} for more information.
}
